Added README file and genomic drawing functions
Added a README file that has some description of some of the functions
offered here. Also added a genomic_drawing.R which provides a small
number of functions that can be used to implement circos synteny plots.
These are not well documented, but I think there may be examples of
usage in the drawing.R file.